+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### Fixed
 
+- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.5.0-alpha.6`
+
 ### Updated
 
 ## [0.5.0-alpha.6] - 2023-03-13